remade checkboxs logic
